+++ lib/src/foundation/bloc.dart
@@ -51,23 +51,26 @@ class Bloc {
   /// This [StreamController] allows to send events of type [EventType].
   /// The [DescribedFeatureOverlay]s will be able to handle these events by checking the
   /// [activeFeatureId] against its own feature id or by considering its current [FeatureOverlayState].
-  final StreamController<EventType> _eventsController =
-      StreamController.broadcast();
+  final _eventsController = StreamController<EventType>.broadcast();
 
   Stream<EventType> get eventsOut => _eventsController.stream;
 
   Sink<EventType> get _eventsIn => _eventsController.sink;
 
   /// The steps consist of the feature ids of the features to be discovered.
-  Iterable<String> _steps;
+  List<String> _steps;
 
   /// Steps that have been previously completed.
   Set<String> _stepsToIgnore;
 
   int _activeStepIndex;
 
-  String get activeFeatureId =>
-      _activeStepIndex == null ? null : _steps?.elementAt(_activeStepIndex);
+  String get activeFeatureId => _steps == null ||
+          _activeStepIndex == null ||
+          _activeStepIndex >= _steps.length ||
+          _activeStepIndex < 0
+      ? null
+      : _steps[_activeStepIndex];
 
   /// This is used to determine if the active feature is already shown by
   /// another [DescribedFeatureOverlay] as [DescribedFeatureOverlay.allowShowingDuplicate]
@@ -105,8 +108,8 @@ class Bloc {
     assert(steps != null && steps.isNotEmpty,
         'You need to pass at least one step to [FeatureDiscovery.discoverFeatures].');
 
-    _steps = steps;
     _stepsToIgnore = await _alreayCompletedSteps;
+    _steps = steps.where((s) => !_stepsToIgnore.contains(s)).toList();
     _activeStepIndex = -1;
 
     await _nextStep();
@@ -114,33 +117,28 @@ class Bloc {
 
   Future<void> completeStep() async {
     if (_steps == null) return;
+    // This will ignore the [onComplete] function of all overlays.
     _eventsIn.add(EventType.complete);
-
-    unawaited(_saveCompletionOf(activeFeatureId));
     await _nextStep();
   }
 
   Future<void> _nextStep() async {
+    if (activeFeatureId != null) unawaited(_saveCompletionOf(activeFeatureId));
     _activeStepIndex++;
     _activeOverlays = 0;
 
     if (_activeStepIndex < _steps.length) {
-      if (_stepsToIgnore.contains(activeFeatureId)) {
-        await _nextStep();
-      } else {
-        _eventsIn.add(EventType.open);
-      }
-      return;
+      _eventsIn.add(EventType.open);
+    } else {
+      // The last step has been completed, so we need to clear the steps.
+      _steps = null;
+      _activeStepIndex = null;
     }
-
-    // The last step has been completed, so we need to clear the steps.
-    _steps = null;
-    _activeStepIndex = null;
   }
 
   void dismiss() {
+    // This will ignore the [onDismiss] function of all overlays.
     _eventsIn.add(EventType.dismiss);
-
     _steps = null;
     _activeStepIndex = null;
     _activeOverlays = 0;